Filtering the Seurat object to reduce size of object

Only keep genes that are part of the interactions db, 2. Only keep cell types with at least N cells (user-defined)

Usage

filtering(seurat_obj, annot, min_cells = 5)

Arguments

seurat_obj

Seurat object

annot

Annotation to use for filtering

min_cells

Minimum number of cells per annotation (default = 5)

genes_oi

list of genes of interest from database

Value

seurat_obj Filtered seurat object

Examples